// src/modules/outpatient/perioperative-nursing-records/models/perioperative-nursing-records-rajal-model.ts
export interface IStatusMental {
    Koma: number;
    Agitasi:number;
    Bingung: number;
    Mengantuk: number;
    Sadar_Penuh: number;
}

export class Status_Mental {
    Koma: number;
    Agitasi:number;
    Bingung: number;
    Mengantuk: number;
    Sadar_Penuh: number;

    constructor(mental: IStatusMental) {
      this.Koma =  mental.Koma;
      this.Agitasi = mental.Agitasi;
      this.Bingung = mental.Bingung;
      this.Mengantuk = mental.Mengantuk;
      this.Sadar_Penuh = mental.Sadar_Penuh;
    }
}

export interface IRiwayatPenyakit {
    Diabetes: number;
    Hepatitis: number;
    Lain_lain: number;
    Tidak_Ada: number;
    Hipertensi: number;
}

export class Riwayat_Penyakit {
    Diabetes: number;
    Hepatitis: number;
    Lain_lain: number;
    Tidak_Ada: number;
    Hipertensi: number;

    constructor(riwayat: IRiwayatPenyakit) {
      this.Diabetes = riwayat.Diabetes;
      this.Hepatitis =  riwayat.Hepatitis;
      this.Lain_lain =  riwayat.Lain_lain;
      this.Tidak_Ada =  riwayat.Tidak_Ada;
      this.Hipertensi =  riwayat.Hipertensi;
    }
}

export interface IPemeriksaanPenunjang {
    Koma: number;
    Rongent: number;
    Biometri: number;
    USG_Mata: number;
    Lain_lain: number;
    Tidak_Ada: number;
    Foto_Fundus: number;
    Laboratorium: number;
}
// ...
export interface IPerioperativeNursingRecordsRajalFormModel {
    Bb: string;
    Rr: string;
    Tb: string;
    Td: string;
    Kgd: string;
    Nadi: string;
    Suhu: string;
    Alergi: string;
    Tanggal: string;
    Alat_Bantu: string;
    ID_Petugas: string;
    Updated_At: string;
    Updated_By: string;
    Skala_Nyeri: string;
    Nama_Petugas: string;
    Status_Mental: IStatusMental;
    Persiapan_Kulit: number;
    Persiapan_Puasa: number;
    Riwayat_Penyakit: IRiwayatPenyakit;
    Alergi_Keterangan: string;
    Lain_Site_Marking: number;
    ID_Perawat_Ruangan: string;
    Operasi_Sebelumnya: string;
    ID_Perawat_Penerima: string;
    Pengobatan_Saat_Ini: string;
    TTD_Perawat_Ruangan: string;
    Nama_Perawat_Ruangan: string;
    Persiapan_Alat_Bantu: number;
    TTD_Perawat_Penerima: string;
    Alat_Bantu_Keterangan: string;
    Nama_Perawat_Penerima: string;
    Operasi_Sebelumnya_Di: string;
    Pemeriksaan_Penunjang: IPemeriksaanPenunjang;
    Lain_Penjelasan_Singkat: number;
    Persiapan_Prothese_Luar: number;
    Persiapan_Prothese_Dalam: number;
    Persiapan_Vaskuler_Akses: number;
    Persiapan_Obat_Disertakan: number;
    Persiapan_Penjepit_Rambut: number;
    Verifikasi_Periksa_Gelang: number;
    Operasi_Sebelumnya_Tanggal: number;
    Persiapan_Kulit_Keterangan: string;
    Persiapan_Puasa_Keterangan: string;
    Riwayat_Penyakit_Keterangan: string;
    Verifikasi_Kelengkapan_X_Ray: number;
    Verifikasi_Periksa_Identitas: number;
    Operasi_Sebelumnya_Keterangan: string;
    Verifikasi_Surat_Izin_Operasi: number,
    Pengobatan_Saat_Ini_Keterangan: string;
    Persiapan_Alat_Bantu_Keterangan: string;
    Verifikasi_Jenis_Lokasi_Operasi: number;
    Verifikasi_Persetujuan_Anestesi: number;
    Pemeriksaan_Penunjang_Keterangan: string;
    Persiapan_Obat_Terakhir_Diberikan: number;
    Persiapan_Prothese_Luar_Keterangan: string;
    Verifikasi_Surat_Pengantar_Operasi: number;
    Persiapan_Prothese_Dalam_Keterangan: string;
    Persiapan_Vaskuler_Akses_Keterangan: string;
    Verifikasi_Kelengkapan_Resume_Medis: number;
    Persiapan_Obat_Disertakan_Keterangan: string;
    Persiapan_Penjepit_Rambut_Keterangan: string;
    Verifikasi_Masalah_Bahasa_Komunikasi: number;
    Verifikasi_Periksa_Gelang_Keterangan: string;
    Verifikasi_Kelengkapan_X_Ray_Keterangan: string;
    Verifikasi_Periksa_Identitas_Keterangan: string;
    Verifikasi_Surat_Izin_Operasi_Keterangan: string;
    Verifikasi_Jenis_Lokasi_Operasi_Keterangan: string;
    Verifikasi_Persetujuan_Anestesi_Keterangan: string;
    Persiapan_Obat_Terakhir_Diberikan_Keterangan: string;
    Verifikasi_Surat_Pengantar_Operasi_Keterangan: string;
    Verifikasi_Kelengkapan_Resume_Medis_Keterangan: string;
    Verifikasi_Masalah_Bahasa_Komunikasi_Keterangan: string;
    Site_Marking: number;
    Site_Marking_Keterangan: string;
    Penjelasan_Singkat: number,
    Penjelasan_Singkat_Keterangan: string,
}
// ...
export interface IPerioperativeNursingRecordsRajalModel extends IDataModel {
    form:  IPerioperativeNursingRecordsRajalFormModel;
    ok: IPerioperativeNursingRecordsFormModel;
    cppt: any;
    asesmen: IAssessmentUgdFormModel;
    ews: INursingEarlyWarning;
}

export class PerioperativeNursingRecordRajalsModel extends DataModel {
    form:  IPerioperativeNursingRecordsRajalFormModel;
    ok: IPerioperativeNursingRecordsFormModel;
    cppt: any;
    asesmen: IAssessmentUgdFormModel;
    ews: INursingEarlyWarning;

    constructor(peri: IPerioperativeNursingRecordsRajalModel) {
      super(peri);
      this.form = peri.form;
      this.ok = peri.ok;
      this.cppt = peri.cppt;
      this.asesmen = peri.asesmen;
      this.ews = peri.ews;
    }
}